Move Transform into Transformer::InMemory
Move spec
