+++ spec/directory_diff/transformer/in_memory_spec.rb
@@ -1,6 +1,6 @@
 require "spec_helper"
 
-describe DirectoryDiff::Transform do
+describe DirectoryDiff::Transformer::InMemory do
   it_behaves_like "a directory transformer" do
     let(:source_directory) { current_directory }
     let(:target_directory) { new_directory }
